Adicionado error handler e index para user e busca por userId

+++ src/main/java/com/user/domain/user/UserRepository.java
@@ -2,8 +2,6 @@ package com.user.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigDecimal;
-
-public interface UserRepository extends JpaRepository<UserEntity, BigDecimal> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
 }
